third_party/WebKit/LayoutTests/media/track/track-disabled.html

<!DOCTYPE html>
<html>
    <head>
        <script src="../media-file.js"></script>
        <script src="../video-test.js"></script>
        <script>
            if (window.testRunner)
                testRunner.dumpAsText();

            waitForEvent('loadstart', function ()
            {
                var v = document.getElementById('vid');
                v.textTracks[0].mode = "disabled";
            });

            waitForEventAndEnd('loadedmetadata');
        </script>
    </head>
    <body>
        <video id="vid" src="../content/test.ogv" autoplay controls>
            <track kind='subtitles' srclang='en' label='English' src='captions-webvtt/captions.vtt' />
        </video>
    </body>
</html>
